Actually made a state machine
the state machine works now. It's made under the scene the_better_player. The scripts are contained in the custom folder of the scripts folder.
Known bugs : state transition to jump or dash cannot happen if the transition to the idle and running is spammed by pressing both the left and right inputs at the same time.