Revert "thrusters_done"
This reverts commit d26031898a7f70ae211a888517887890d2fd4a00.

Reason: The branch in which this code was written was behind the branch it was supposed to be written in by about 2 commits. Also would be easier to copy useful snippets from the website and paste it in this branch than trying to rework the implementation from this commit.
Like seriously, there is an entire resource for export variables, why paste them into the controller code. could've just asked